PromptEngineering/LLM_tests_framework/src/frameworks/integration_tracker.py
from typing import Dict, List, Optional
import asyncio
from dataclasses import dataclass
import numpy as np
from .performance_tracker import PerformanceTracker

@dataclass
class TestConfig:
    model_configs: Dict[str, Dict]
    rag_configs: Dict[str, Dict]
    prompt_templates: Dict[str, str]
    test_scenarios: List[Dict]

class IntegrationTracker:
    def __init__(self, config: TestConfig):
        self.config = config
        self.results_history = []
        self.performance_tracker = PerformanceTracker()
        
    async def run_comprehensive_tests(self):
        """Run all tests"""
        tasks = [
            self._test_base_models(),
            self._test_rag_implementations(),
            self._test_prompt_variations(),
            self._test_integration_scenarios()
        ]
        results = await asyncio.gather(*tasks)
        return self._aggregate_results(results)
    
    async def _test_base_models(self):
        """Test basic model capabilities"""
        results = {}
        for model_name, model_config in self.config.model_configs.items():
            results[model_name] = await self._evaluate_model(model_name, model_config)
        return results
    
    async def _test_rag_implementations(self):
        """Test different RAG implementations"""
        results = {}
        for rag_name, rag_config in self.config.rag_configs.items():
            results[rag_name] = await self._evaluate_rag(rag_name, rag_config)
        return results
    
    async def _test_prompt_variations(self):
        """Test different prompt variations"""
        results = {}
        for template_name, template in self.config.prompt_templates.items():
            results[template_name] = await self._evaluate_prompt_template(
                template_name, template
            )
        return results
    
    async def _test_integration_scenarios(self):
        """Test integration scenarios"""
        results = {}
        for scenario in self.config.test_scenarios:
            results[scenario['name']] = await self._evaluate_scenario(scenario)
        return results
    
    def _aggregate_results(self, results: List[Dict]) -> Dict:
        """Aggregate all test results"""
        return {
            'base_models': results[0],
            'rag_implementations': results[1],
            'prompt_variations': results[2],
            'integration_scenarios': results[3],
            'summary': self._generate_summary(results),
            'recommendations': self._generate_recommendations(results)
        }
    
    def _generate_summary(self, results: List[Dict]) -> Dict:
        """Generate summary of results"""
        return {
            'best_performing_combination': self._find_best_combination(results),
            'performance_metrics': self.performance_tracker._calculate_performance_metrics(),
            'cost_analysis': self.performance_tracker.analyze_costs(),
            'latency_analysis': self.performance_tracker._analyze_latency()
        }
    
    def _find_best_combination(self, results: List[Dict]) -> Dict:
        """Find the best combination of components"""
        combinations = self._generate_combinations(results)
        scored_combinations = [
            (combo, self._score_combination(combo))
            for combo in combinations
        ]
        return max(scored_combinations, key=lambda x: x[1])[0]

    # Placeholder methods for evaluation
    async def _evaluate_model(self, model_name: str, model_config: Dict) -> float:
        pass

    async def _evaluate_rag(self, rag_name: str, rag_config: Dict) -> float:
        pass

    async def _evaluate_prompt_template(self, template_name: str, template: str) -> float:
        pass

    async def _evaluate_scenario(self, scenario: Dict) -> float:
        pass

    def _generate_combinations(self, results: List[Dict]) -> List[Dict]:
        pass

    def _score_combination(self, combination: Dict) -> float:
        pass

    def _generate_recommendations(self, results: List[Dict]) -> List[str]:
        pass